requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
weaviate-client>=3.24.1
openai>=1.3.0
asyncio>=3.4.3